# documentation/swagger-docs/backoffice-service-swagger.json
{"openapi":"3.0.0","info":{"title":"backoffice-service service","description":"The backoffice-service service API description","version":"1.0","contact":{}},"tags":[{"name":"backoffice-service","description":""}],"servers":[],"components":{"schemas":{"ChatRequest":{"type":"object","properties":{"idConversacion":{"type":"string","description":"ID de la conversación","example":"234234"},"pushName":{"type":"string","description":"Nombre del usuario","example":"Gibran Garcia"},"q":{"type":"string","description":"Consulta de búsqueda","example":"sabes de matemáticas"},"tosearch":{"description":"Campos a buscar","example":["*"],"type":"array","items":{"type":"string"}},"max":{"type":"number","description":"Número máximo de resultados","example":20}},"required":["idConversacion","pushName","q","tosearch","max"]},"DashboardMetricsDto":{"type":"object","properties":{"totalUsers":{"type":"number","description":"Número total de usuarios","example":100},"totalConversaciones":{"type":"number","description":"Número total de conversaciones","example":500}},"required":["totalUsers","totalConversaciones"]},"RecentActivityDto":{"type":"object","properties":{"user":{"type":"string","description":"Nombre del usuario","example":"John Doe"},"idConversacion":{"type":"string","description":"ID único de la conversación","example":"1234567890"},"timestamp":{"type":"number","description":"Marca de tiempo de la actividad reciente ","example":"01/4/2025 10:59 PM"}},"required":["user","idConversacion","timestamp"]},"DashboardResponseDto":{"type":"object","properties":{"metrics":{"description":"Métricas generales del dashboard","allOf":[{"$ref":"#/components/schemas/DashboardMetricsDto"}]},"recentActivity":{"description":"Actividad reciente del dashboard","type":"array","items":{"$ref":"#/components/schemas/RecentActivityDto"}}},"required":["metrics","recentActivity"]},"UsersDto":{"type":"object","properties":{}}}},"paths":{"/api/v2/conversacion/chat":{"post":{"operationId":"getChat","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}}},"responses":{"201":{"description":""}},"tags":["backoffice-service"]}},"/api/dashboard/metrics":{"get":{"operationId":"getDashboardMetrics","summary":"Obtener métricas del dashboard","description":"Este endpoint devuelve las métricas generales y la actividad reciente del dashboard.","parameters":[],"responses":{"200":{"description":"Métricas obtenidas exitosamente.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardResponseDto"}}}},"500":{"description":"Error interno del servidor."}},"tags":["backoffice-service"]}},"/api/dashboard/filtered":{"get":{"operationId":"getFilteredConversations","summary":"Obtener conversaciones filtradas por fecha","description":"Este endpoint devuelve las conversaciones filtradas por un rango de fechas.","parameters":[{"name":"startDate","required":true,"in":"query","description":"Fecha de inicio en formato DD-MM-YYYY","example":"01-04-2025","schema":{"type":"string"}},{"name":"endDate","required":true,"in":"query","description":"Fecha de fin en formato DD-MM-YYYY","example":"02-04-2025","schema":{"type":"string"}}],"responses":{"200":{"description":"Conversaciones filtradas obtenidas exitosamente."},"400":{"description":"Error en la solicitud. Asegúrate de que las fechas sean válidas."},"500":{"description":"Error interno del servidor."}},"tags":["backoffice-service"]}},"/api/dashboard/filtered/download":{"get":{"operationId":"downloadFilteredConversationsExcel","summary":"Descargar conversaciones filtradas en formato Excel","description":"Este endpoint genera y descarga un archivo Excel con las conversaciones filtradas por un rango de fechas.","parameters":[{"name":"startDate","required":true,"in":"query","description":"Fecha de inicio en formato DD-MM-YYYY","example":"01-04-2025","schema":{"type":"string"}},{"name":"endDate","required":true,"in":"query","description":"Fecha de fin en formato DD-MM-YYYY","example":"02-04-2025","schema":{"type":"string"}}],"responses":{"200":{"description":"Archivo Excel generado y descargado exitosamente."},"400":{"description":"Error en la solicitud. Asegúrate de que las fechas sean válidas."},"500":{"description":"Error interno del servidor."}},"tags":["backoffice-service"]}},"/auth/login":{"post":{"operationId":"login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsersDto"}}}},"responses":{"201":{"description":""}}}}}}